Fixed clang compilation

+++ src/cam_exception.cc
@@ -98,7 +98,7 @@ namespace cam
 
     void operator<<= (cxxtools::SerializationInfo& si, const CamException & exception)
     {
-        si.addMember("errorCode") <<= exception.m_code;
+        si.addMember("errorCode") <<= uint8_t(exception.m_code);
         si.addMember("whatArg") <<= exception.m_whatArg;
         exception.fillSerializationInfo(si.addMember("extraData"));
     }
